Module sc_light::blockchain

Light client blockchain backend. Only stores headers and justifications of recent blocks. CHT roots are stored for headers of ancient blocks.

Modules

cht
Canonical hash trie definitions and helper functions.

well_known_cache_keys
A list of all well known keys in the blockchain cache.

Structs

Blockchain
Light client blockchain.

BlockchainInfo
Blockchain info

Enums

BlockStatus
Block status.

LocalOrRemote
Remote header.

NewBlockState
State of a new block.

Traits

AuxStore
Provides access to an auxiliary database.

BlockchainBackend
Blockchain database backend. Does not perform any validation.

BlockchainCache
Blockchain optional data cache.

BlockchainHeaderBackend
Blockchain database header backend. Does not perform any validation.

ProvideCache
Provides access to the optional cache.

ProvideChtRoots
Provide CHT roots. These are stored on a light client and generated dynamically on a full client.

RemoteBlockchain
Futures-based blockchain backend that either resolves blockchain data locally, or fetches required data from remote node.

Storage
Light client blockchain storage.