Module sc_light::call_executor

Methods that light client could use to execute runtime calls.

Structs

GenesisCallExecutor
Call executor that is able to execute calls only on genesis state.

Functions

check_execution_proof
Check remote contextual execution proof using given backend.

check_execution_proof_with_make_header
Check remote contextual execution proof using given backend and header factory.

prove_execution
Prove contextual execution using given block header in environment.