Function sc_light::call_executor::prove_execution

pub fn prove_execution<Block, S, E>(
    state: S, 
    header: Block::Header, 
    executor: &E, 
    method: &str, 
    call_data: &[u8]
) -> ClientResult<(Vec<u8>, StorageProof)> where
    Block: BlockT,
    S: StateBackend<HashFor<Block>>,
    E: CallExecutor<Block>, 

Prove contextual execution using given block header in environment.

Method is executed using passed header as environment’ current block. Proof includes both environment preparation proof and method execution proof.