Module sc_light::fetcher

Light client data fetcher. Fetches requested data from remote full nodes.

Modules

cht

Canonical hash trie definitions and helper functions.

Structs

ChangesProof

Key changes read proof.

LightDataChecker

Remote data checker.

RemoteBodyRequest

Remote block body request

RemoteCallRequest

Remote call request.

RemoteChangesRequest

Remote key changes read request.

RemoteHeaderRequest

Remote canonical header request.

RemoteReadChildRequest

Remote storage read child request.

RemoteReadRequest

Remote storage read request.

StorageProof

A proof that some set of key-value pairs are included in the storage trie. The proof contains the storage values so that the partial storage backend can be reconstructed by a verifier that does not already have access to the key-value pairs.

Traits

BlockchainStorage

Light client blockchain storage.

FetchChecker

Light client remote data checker.

Fetcher

Light client data fetcher. Implementations of this trait must check if remote data is correct (see FetchedDataChecker) and return already checked data.