Crate sc_light

Light client components.

Re-exports

pub use backend::*;

pub use blockchain::*;

pub use call_executor::*;

pub use fetcher::*;

Modules

backend
Light client backend. Only stores headers and justifications of blocks. Everything else is requested from full nodes on demand.

blockchain
Light client blockchain backend. Only stores headers and justifications of recent blocks. CHT roots are stored for headers of ancient blocks.

call_executor
Methods that light client could use to execute runtime calls.

fetcher
Light client data fetcher. Fetches requested data from remote full nodes.

Functions

new_fetch_checker
Create an instance of fetch data checker.

new_light_backend
Create an instance of light client backend.

new_light_blockchain
Create an instance of light client blockchain backend.